requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio>=3.0.0
soundfile>=0.12.1
numpy>=1.24.0
pydub>=0.25.1
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.7
